+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.21.0
 
 * Removed "style-disabled" property and GetDisabledStyle function
+* Added GoogleFonts field to AppParams
 
 # v0.20.0